requirements.txt
pandas
openpyxl
odfpy
llama-cpp-python
requests
py2gift
